[settings] Add output_buildfile_suffix option
This option allows the user to set any suffix to the build files
Default is BUILD
Other values can be for example: BUILD.bazel
Any value is accepted.

This is to allow raze to follow the default behaviour of other bazel
related tools that generate BUILD.bazel instead of BUILD firstly to
separate what is generated and what is written and to avoid name
collisions on case insensitive file systems (Windows + WSL) since some
crates can have a directory named build which leads to cargo-raze
unable to create a file called BUILD. A project having a directory
named BUILD.bazel is much less likely or less to none chance.