[BUILD] Fix inconsistent casing in targets and dep targets
Inconsistent casing where dep targets were slugified and target names
were not caused errors when crate names were not all lowercase.
This fixes the issue by removing the unnecessary slugifying from the
target name (but keeps it in versioning)
An example for this issue is with security-framework-sys which depends
on the MacTypes-sys crate.
To test this, I've added security-framework-sys as a dependency to both
remote and vendored complicated_cargo_library