[Planner] Allow for renamed or aliased crates
This turns out to be important for some fairly important crates.

For example, consider the `futures-0.3.x` crates. This crate offers a
feature (`compat`) that allows for the interaction with `futures01`
codebases.

Internally this is accomplished with the following dependency
declaration

```toml
[dependencies]
futures_01 = { version = "0.1", package = "futures" }
```

This makes the crate futures-0.1.x available as a dependency to
futures-0.3.x, but named in compilation as `futures_01`.
The compilation environment uses the `--extern` flags for rustc to
finish this remapping.

This is exposed to rules_rust as an `alias` attribute, allowing
rules_rust to honor these renames.