Don't copy Cargo.toml to tempdir where possible
I'm not entirely clear on why we're copying out in the first place, but
I added a comment with my best guess.

Avoid doing this where possible, because it breaks support for Cargo
workspaces.

Alternatively, we could try to parse the Cargo.toml for its workspace
members and copy those too, but that seems to be breaking more
abstraction layers.