Added additional smoketests and improved testing speed (#224)
This change improves the speed of tests by combining all build steps into 1 command and avoids having to startup Bazel for each set of tests.

There is also now a test for a project that has no dependencies. This tests that raze doesn't generate broken files in this case

Finally, this change adds a launch_test macro which is used to create tests for each example binary and launch it as part of a test. This macro was created to try and keep the examples as straight forward as possible (not having to dump some test code in there). This should make it easier to catch regressions.