Allow matching triples directly (#232)
Before this change, the get_matching_bazel_triples would only allow matching cfg(...) targets. It seems that the intended behaviour was to allow passing triples (such as `x86_64-unknown-linux-gnu`) directly, as per the `format!("cfg(target = \"{}\")", target)` replacement at the beginning of the function. Unfortunately, the match was missing the right case below (`Predicate::Target(tp)` only matches builtin `target_{arch,os,...}`).

This makes this function more consistent with `is_bazel_supported_platform` above, and it seems to solve a bug where target specific dependencies specified with triples (such as is the case for the `winapi` crate in one of the `complicated_cargo_library` example) would be ignored.