platform-specific aliases were missing (#244)
* platform-specific aliases were missing

* avoid extra newline when aliases is omitted

* update examples via smoke_test

The length check in common_attrs is only checking if there are any
targeted deps, so targeted deps without any aliases end up with an
empty aliases entry instead of it being excluded completely. What we
really want to do if check if deps.aliased_dependencies is non-empty
for any of the provided targeted deps, but I'm not sure how/if that
can be done in Tera.