Set links attribute
This was added to rules_rust in
https://github.com/bazelbuild/rules_rust/pull/480 and we started parsing
out the metadata in cargo_raze in #281 but didn't start populating it in
generated BUILD files.

This enables us to remove a hack in rules_rust where we try to infer the
correct value for the `links` attribute if it wasn't set, based on the
name, which in turn will allow us to remove the `crate_name` attribute
completely.