Support additional registries (#412)
This relies on the registry information being in .cargo/config.toml.

The reason for this is that we need to run `cargo metadata` before we
even parse out config, and we can't run `cargo metadata` without having
this data active for that `cargo` invocation. It also simplifies
implementation a lot, by just copying around one file rather than
needing to generate config for `cargo` to use.