Fix features selection, broken by #478
It works properly with no explicit targets set (aka "use all the
targets"), and it also works properly with some of the examples that set
`default-members = []`.

I think there's still something wrong, because it dropped features from
some Windows-specific crates relative to before #478, but I'm not sure
what's supposed to happen. At least all the tests in this repository
pass on Linux again...