Document the commands to regenerate checked-in outputs
Looking through the commit history, it looks like forgetting to
re-generated some of the outputs that get checked has happened multiple
times. Doing this often makes bugs in new changes obvious when things
fail to build.

Ideally CI would verify this (there's a TODO for it in presubmit.yml),
but as a starting point let's at least document it.